Consolidate Configuration and ConfigurationAttributes
