+++ lib/gemfury/configuration.rb
@@ -1,36 +1,48 @@
-require 'gemfury/configuration_attributes'
-
 module Gemfury
-  # Defines constants and methods related to configuration
   module Configuration
-    include Gemfury::ConfigurationAttributes
 
-    # An array of valid keys in the options hash when configuring
-    VALID_OPTIONS_KEYS = CONFIGURATION_DEFAULTS.keys.freeze
+    CONFIGURATION_DEFAULTS = {
+      :user_api_key => nil,
+      :adapter => :net_http,
+      :endpoint => 'https://api.fury.io/',
+      :gitpoint => 'https://git.fury.io/',
+      :pushpoint => 'https://push.fury.io/',
+      :user_agent => "Gemfury RubyGem #{Gemfury::VERSION}",
+      :api_version => 1,
+      :account => nil
+    }.freeze
 
-    # When this module is extended, set all configuration options to their default values
-    def self.extended(base)
-      base.reset
-    end
+    # user API key, also known as "full access token"
+    # @return [String]
+    attr_accessor :user_api_key
 
-    # Convenience method to allow configuration options to be set in a block
-    def configure
-      yield self
-    end
+    # The adapter that will be used to connect
+    # @return [Symbol]
+    attr_accessor :adapter
 
-    # Create a hash of options and their values
-    # @return [Hash] the options and their values
-    def options
-      options = {}
-      VALID_OPTIONS_KEYS.each{|k| options[k] = send(k)}
-      options
-    end
+    # The endpoint that will be used to connect
+    # @return [String]
+    attr_accessor :endpoint
 
-    # Reset all configuration options to defaults
-    # @return [Configuration] The default configuration
-    def reset
-      CONFIGURATION_DEFAULTS.each { |k, v| send("#{k}=", v) }
-      self
-    end
+    # The HTTP endpoint for git repo (used for .netrc credentials)
+    # @return [String]
+    attr_accessor :gitpoint
+
+    # The endpoint for the Push API
+    # @return [String]
+    attr_accessor :pushpoint
+
+    # The value sent in the 'User-Agent' header
+    # @return [String]
+    attr_accessor :user_agent
+
+    # Gemfury remote API version
+    # @return [Integer]
+    attr_accessor :api_version
+
+    # The account to impersonate, if you have permissions for multiple accounts
+    # (If nil, no impersonation)
+    # @return [String]
+    attr_accessor :account
   end
 end
